+++ vignettes/EHRtemporalVariability.R
@@ -1,32 +1,32 @@
-## ----setup, include=FALSE--------------------------------------------------
-knitr::opts_chunk$set(echo = TRUE)
+## ----setup, include=FALSE-----------------------------------------------------
+knitr::opts_chunk$set(echo = TRUE, warning=FALSE)
 
-## ----cran, message=FALSE, eval=FALSE, warning=FALSE------------------------
+## ----cran, message=FALSE, eval=FALSE, warning=FALSE---------------------------
 #  install.packages("EHRtemporalVariability")
 #  library(EHRtemporalVariability)
 
-## ----crantrue, echo=FALSE, message=FALSE, eval=TRUE, warning=FALSE---------
+## ----crantrue, echo=FALSE, message=FALSE, eval=TRUE, warning=FALSE------------
 library(EHRtemporalVariability)
 
-## ----devtls, message=FALSE, eval=FALSE, warning=FALSE----------------------
+## ----devtls, message=FALSE, eval=FALSE, warning=FALSE-------------------------
 #  install.packages("devtools")
 #  library(devtools)
 
-## ----bioC, message=FALSE, eval=FALSE, warning=FALSE------------------------
-#  install_github( "hms-dbmi/EHRtemporalVariability" )
+## ----bioC, message=FALSE, eval=FALSE, warning=FALSE---------------------------
+#  install_github("hms-dbmi/EHRtemporalVariability")
 #  library( EHRtemporalVariability )
 
-## ----EHRtemporalVariabilityObj1, echo = FALSE, eval = TRUE-----------------
+## ----EHRtemporalVariabilityObj1, echo = FALSE, eval = TRUE--------------------
 githubURL <- "http://github.com/hms-dbmi/EHRtemporalVariability-DataExamples/raw/master/variabilityDemoNHDS.RData"
 load(url(githubURL))
 
-## ----variabilityObj2, eval=TRUE--------------------------------------------
+## ----variabilityObj2, eval=TRUE-----------------------------------------------
 class( probMaps$`diagcode1-phewascode` )
 
-## ----igtProjObj, eval=TRUE-------------------------------------------------
+## ----igtProjObj, eval=TRUE----------------------------------------------------
 class( igtProjs$`diagcode1-phewascode` )
 
-## ----readCSV, eval=TRUE----------------------------------------------------
+## ----readCSV, eval=TRUE-------------------------------------------------------
 dataset <- read.csv2( "http://github.com/hms-dbmi/EHRtemporalVariability-DataExamples/raw/master/nhdsSubset.csv", 
                       sep  = ",",
                       header = TRUE, 
@@ -35,7 +35,7 @@ dataset <- read.csv2( "http://github.com/hms-dbmi/EHRtemporalVariability-DataExa
                                       "numeric" , rep( "factor", 22 ) ) )
 head( dataset)
 
-## ----formatDate2, eval=TRUE------------------------------------------------
+## ----formatDate2, eval=TRUE---------------------------------------------------
 class( dataset$date )
 datasetFormatted <- EHRtemporalVariability::formatDate(
               input         = dataset,
@@ -45,7 +45,7 @@ datasetFormatted <- EHRtemporalVariability::formatDate(
 head( datasetFormatted )[1:5, 1:5]
 class( datasetFormatted$date )
 
-## ----icd9toPheWAS, eval=TRUE, message=FALSE, warning=FALSE-----------------
+## ----icd9toPheWAS, eval=TRUE, message=FALSE, warning=FALSE--------------------
 datasetPheWAS <- icd9toPheWAS(data           = datasetFormatted,
                               icd9ColumnName = "diagcode1",
                               phecodeDescription = TRUE,
@@ -55,16 +55,16 @@ datasetPheWAS <- icd9toPheWAS(data           = datasetFormatted,
 
 head( datasetPheWAS[, c( "diagcode1", "diagcode1-phewascode")] )
 
-## ----estimateDataTemporalMap, eval=FALSE-----------------------------------
+## ----estimateDataTemporalMap, eval=FALSE--------------------------------------
 #  probMaps <- estimateDataTemporalMap(data           = datasetPheWAS,
 #                                      dateColumnName = "date",
 #                                      period         = "month")
 
-## ----estimateDataTemporalMapOutput, eval=TRUE------------------------------
+## ----estimateDataTemporalMapOutput, eval=TRUE---------------------------------
 class( probMaps )
 class( probMaps[[ 1 ]] )
 
-## ----estimateDataTemporalMapSupport, eval=FALSE----------------------------
+## ----estimateDataTemporalMapSupport, eval=FALSE-------------------------------
 #  supports <- vector("list",2)
 #  names(supports) <- c("age","diagcode1")
 #  supports[[1]] <- 1:18
@@ -74,7 +74,7 @@ class( probMaps[[ 1 ]] )
 #                                      period         = "month",
 #                                      supports       = supports)
 
-## ----trimDataTemporalMap, eval=TRUE----------------------------------------
+## ----trimDataTemporalMap, eval=TRUE-------------------------------------------
 class( probMaps[[1]] )
 probMapTrimmed <- trimDataTemporalMap( 
                         dataTemporalMap = probMaps[[1]],
@@ -83,39 +83,68 @@ probMapTrimmed <- trimDataTemporalMap(
                                       )
 class( probMapTrimmed )
 
-## ----estimateIGTProjection, eval=FALSE-------------------------------------
-#  igtProj <- estimateIGTProjection( dataTemporalMap = probMaps[[1]],
-#                                    dimensions      = 3,
-#                                    startDate       = "2000-01-01",
-#                                    endDate         = "2010-12-31")
+## ----estimateIGTProjection, eval=TRUE-----------------------------------------
+igtProj <- estimateIGTProjection( dataTemporalMap = probMaps[[1]], 
+                                  dimensions      = 2, 
+                                  startDate       = "2000-01-01", 
+                                  endDate         = "2010-12-31")
 
-## ----estimateIGTProjectionOutput, eval=FALSE-------------------------------
-#  class( igtProj )
+## ----estimateIGTProjectionOutput, eval=TRUE-----------------------------------
+class( igtProj )
 
-## ----sapplyestimateIGTProjection, eval=FALSE-------------------------------
+## ----sapplyestimateIGTProjection, eval=FALSE----------------------------------
 #  igtProjs <- sapply ( probMaps, estimateIGTProjection )
 #  names( igtProjs ) <- names( probMaps )
-#  class( igtProj[[ 1 ]] )
 
-## ----loadExampleFile, eval=TRUE--------------------------------------------
+## ----loadExampleFile, eval=TRUE-----------------------------------------------
 githubURL <- "https://github.com/hms-dbmi/EHRtemporalVariability-DataExamples/raw/master/variabilityDemoNHDS.RData"
 load(url(githubURL))
 
-## ----plotHeatmap, eval=TRUE------------------------------------------------
+## ----plotHeatmap, eval=TRUE---------------------------------------------------
 plotDataTemporalMap(
     dataTemporalMap =  probMaps[["diagcode1-phewascode"]],
     startValue = 2,
     endValue = 20,
     colorPalette    = "Spectral")
 
-## ----plotIGTprojection, eval=TRUE------------------------------------------
+## ----plotIGTprojection, eval=TRUE---------------------------------------------
 plotIGTProjection( 
     igtProjection   =  igtProjs[["diagcode1-phewascode"]],
     colorPalette    = "Spectral", 
     dimensions      = 2)
 
-## ----saveRData, eval=FALSE-------------------------------------------------
+## ----plotIGTprojectionTrajectory, eval=TRUE-----------------------------------
+plotIGTProjection( 
+    igtProjection   =  igtProjs[["diagcode1-phewascode"]],
+    colorPalette    = "Spectral", 
+    dimensions      = 2,
+    trajectory      = TRUE)
+
+## ----saveRData, eval=FALSE----------------------------------------------------
 #  names( probMaps )
 #  names( igtProjs )
 #  save(probMaps, igtProjs, file = "myExport.RData")
+
+## ----dbscan, message=FALSE, eval=FALSE, warning=FALSE-------------------------
+#  install.packages("dbscan")
+#  library(dbscan)
+
+## ----dbscantrue, echo=FALSE, message=FALSE, eval=TRUE, warning=FALSE----------
+library(dbscan)
+
+## ----temporalSubgroupsClustering, eval=TRUE-----------------------------------
+# We set the minimum number of batches in a subgroup as 2 
+# We set eps based on the knee of the following KNNdistplot, at around 0.023
+# kNNdistplot(igtProj@projection, k = 2, all = FALSE)
+igtProj = igtProjs[["diagcode1-phewascode"]]
+# We select the 2 first dimensions for consistency with the IGT plot examples above
+dbscanResults <- dbscan(igtProj@projection[,c(1,2)], eps = 0.023, minPts = 2)
+clusterNames  <- vector(mode = "character", length = 10)
+clusterNames[dbscanResults$cluster == 0] <- "Outlier batches"
+clusterNames[! dbscanResults$cluster == 0] <- paste("Temporal subgroup",dbscanResults$cluster[! dbscanResults$cluster == 0])
+plotly::plot_ly(x = igtProj@projection[,1], y = igtProj@projection[,2],
+              color = as.factor(clusterNames),
+              type = "scatter", mode = "markers",
+              text = paste0("Date: ",igtProj@dataTemporalMap@dates)) %>%
+              plotly::config(displaylogo = FALSE)